[FIO toup] sks ta: duplicate ref tee attribute buffer for mem alignment
Duplicate attribute reference buffer when loading TEE attributes in
order to make sure the memory buffer is properly aligned. This is needed
as there is no alignment guarantee in the original SKS buffer list as it
is serialized and the order can change when adding/removing attributes.

The duplicated buffer is only needed up to TEE_PopulateTransientObject,
as at that point the TEE object is properly set and the ref buffer can
be released.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>