[FIO internal] caam: workaround CAAM locked by HAB during SDP boot
On the imx7ulp and during SDP boot the CAAM is not being unlocked
despite the request being made in the signed image (CSF section).

When this is the case, and if the CAAM is being locked with invalid
MID configs, the CAAM driver would fail to initialize as the MID
registers can not be reconfigured. And this would panic booting the
trust zone.

Ignoring this error and allowing the system to continue booting seems
to fulfill the software upgrade use case requirement (which is where we
use SDP for).

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>