[FIO fromlist] drivers: imx_i2c: add support for imx6ull
Support native I2C access on imx6ull (this SoC has an additional
register to configure the daisy chain in the iomuxc).

The driver assumes that the bootloader configured the peripheral clock
to run out of the 24MHz oscillator (as the linux kernel does and as
recommended by the I2C errata [1]).

[1] imx6ull errata, Rev 2, 10/2019, ERR007805
Tested on imx6ull-evk 14x14 with the bus at 100Kbps.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>