[FIO internal] fiovb: add support to delete values
Add support to delete values (objects).

Variables (tee objects) are allowed to be created with write, but it is
not currently possible to delete the object completely from tee after
it gets created.

This will also allow vendor specific values to be created as write-only,
which can only be updated if deleted and created from scratch.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>